requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
mysql-connector-python==9.4.0
sqlparse==0.5.3
typing_extensions==4.14.1
python-decouple==3.8
openpyxl==3.1.5
